:hammer: Extract utility functions
